Fix runtime warnings abut changedName(QString) being redefined
A runtime warning is issued for every instance of a BeerXMLElement subclass that redefines changedName(QString) in its header, even though this is already defined in the BeerXMLElement header file. AFAIK, there is no reason to redeclare this method, and removing the redeclarations gets rid of the warnings without any apparent ill effects.